Fix controller reference
